fix order

+++ src/main/java/t1708m/fashion/repository/OrderDetailRepository.java
@@ -4,5 +4,5 @@ package t1708m.fashion.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import t1708m.fashion.entity.OrderDetail;
 
-public interface OrderDetailRepository  extends JpaRepository<OrderDetail, Integer> {
+public interface OrderDetailRepository  extends JpaRepository<OrderDetail, Long> {
 }
